Update docs, pin REST API requirements, use modelId for LiteLLM
These changes address pul request feedback for adding comments for inter-file dependencies and
updates to the README file. This also modifies the LiteLLM configuration to use the user-provided
modelId if it exists, otherwise default back to the model name. This allows for users to use the
same model and weights, but using two different containers.