Add LiteLLM router to FastAPI container (#14)
* Add LiteLLM router to FastAPI container

* Add LiteLLM config to test config.yaml

* Update docs, pin REST API requirements, use modelId for LiteLLM

These changes address pull request feedback for adding comments for inter-file dependencies and
updates to the README file. This also modifies the LiteLLM configuration to use the user-provided
modelId if it exists, otherwise default back to the model name. This allows for users to use the
same model and weights, but using two different containers.

* Add dependency missed when pinning versions

---------

Signed-off-by: Peter Muller <petermul@amazon.com>