Add PostgreSQL db for LiteLLM scaling (#45)
* Add PostgreSQL db for LiteLLM scaling

This set of changes adds a PostgreSQL database to the Serve infrastructure to help
with scaling LiteLLM across multiple instances. This database will be used between
multiple REST API containers so that models that are dynamically added to a running
LiteLLM instance will update the DB and be available across all other REST API
containers behind the LISA Serve endpoint.

Tests have increased error counts because the database is a Single-AZ db. This is
expected to be fine because the database is only used for this one purpose and is
not anticipated to have any form of high load on it.

This database will be used in further model management feature development.

* Add configuration option for database port