Lock down LiteLLM administration APIs to admin users
This set of changes refactors the JWT parsing so that functions can be reused
to also validate that a user is an admin user or not. Only admin users should
be allowed to perform LiteLLM administration actions, like adding or removing
models, so these changes help to enforce that at the LISA REST API level.