Add GetModel and DeleteModel API handlers
This set of changes implements the initial versions of the GetModel and DeleteModel
management APIs. The DeleteModel will still have more work once there is a
CloudFormation stack to delete. With these changes, we now have the ability
to get the metadata of a specific model, and then delete that model from LiteLLM
if desired.