Add DeleteModel state machine for model management
This set of changes adds a state machine for handling the deletion of a model
to serve as a framework for other state machines that will be needed for CreateModel
or UpdateModel. This includes, but does not implement files for CreateModel, but
those will be filled out in a future revision.