Add Create and Delete state machines for model management (#64)
* Add DeleteModel state machine for model management

This set of changes adds a state machine for handling the deletion of a model
to serve as a framework for other state machines that will be needed for CreateModel
or UpdateModel. This includes, but does not implement files for CreateModel, but
those will be filled out in a future revision.

* Add CreateModel state machine skeleton

Adds Lambda functions and state machine to support creating a model. Lambda
functions must be filled in, but the state machine architecture should stay
relatively the same.

* Create state machine Lambdas in VPC

* Invoke delete workflow from lambda handler, fix formatting issues

* Add handler and state machine invocation for CreateModel

* Formatting fixes