Fix: remove import/export dependency between stacks on management key
In 3.0.0 we introduced a secret key to allow for Lambda functions to make
LiteLLM management calls while not explicitly being an admin user from the
user's IdP. The way how we referenced it across different stacks caused
an import/export dependency between stacks, which made the affected resources
be immutable, which cannot happen for the sake of future CDK deployments.
This set of changes removes the explicit dependency in favor of an SSM
StringParameter reference, which lets us take on the dependency without
CloudFormation having immutability issues with it.