Add UpdateModel workflow, remove Start and Stop API, minor refactor (#106)
* Add UpdateModel workflow, remove Start and Stop API, minor refactor

This set of changes adds a stubbed state machine for implementing the UpdateModel workflow,
and at the same time removes the unused Start and Stop APIs. Because LiteLLM is now only
referenced inside the state machines, and not in the API handler, the LiteLLM client was
also removed from the BaseHandler class, so it was no longer needed in the API definitions
file. In addition, the LiteLLM client was the only reason we needed the Request object from
FastAPI, so that has been removed from other API definitions as well.

* Update comment accuracy